controller updated for add vehicle.

+++ zippyusedautos/admin/index.php
@@ -8,10 +8,24 @@ require_once('../model/database.php');
 require_once('../model/vehicles_db.php');
 
 //Calls the delete funciton if a POST was submitted.
+
 if ($_SERVER['REQUEST_METHOD'] == 'POST') {
-    $vehicle_id = $_POST['vehicle_id'];
-    delete_vehicle($db, $vehicle_id);
+    if ($_POST['action'] == 'delete') {
+        $vehicle_id = $_POST['vehicle_id'];
+        delete_vehicle($db, $vehicle_id);
+
+    } else if ($_POST['action'] == 'add') {
+        $year = $_POST['year'];
+        $model = $_POST['model'];
+        $price = $_POST['price'];
+        $make_id = $_POST['make_id'];
+        $type_id = $_POST['type_id'];
+        $class_id = $_POST['class_id'];
+        add_vehicle($db, $year, $model, $price, $make_id, $type_id, $class_id);
+    }
 }
+
+
 //Loads the list of vehicles. 
 $vehicles = get_vehicles($db, $order);
